+++ cvxpygen/__init__.py
@@ -1,0 +1,1 @@
+from cvxpygen.generator import Generator
