+++ cvxpygen/solvers/__init__.py
@@ -1,0 +1,7 @@
+from cvxpygen.solvers.interface import SolverInterface, QPCanonMixin, UNSUPPORTED_ON_WINDOWS
+from cvxpygen.solvers.osqp import OSQPInterface
+from cvxpygen.solvers.scs import SCSInterface
+from cvxpygen.solvers.ecos import ECOSInterface
+from cvxpygen.solvers.clarabel import ClarabelInterface
+from cvxpygen.solvers.qoco import QOCOGENInterface, QOCOInterface
+from cvxpygen.solvers.pdaqp import PDAQPInterface
